Modify for .net6.0

+++ src/WFLite/Activities/IO/FileDeleteActivity.cs
@@ -9,13 +9,14 @@
 
 using System.IO;
 using WFLite.Bases;
+using WFLite.Extensions;
 using WFLite.Interfaces;
 
 namespace WFLite.Activities.IO
 {
     public class FileDeleteActivity : SyncActivity
     {
-        public IOutVariable<string> Path
+        public IOutVariable<string>? Path
         {
             private get;
             set;
@@ -25,14 +26,23 @@ namespace WFLite.Activities.IO
         {
         }
 
-        public FileDeleteActivity(IOutVariable<string> path)
+        public FileDeleteActivity(IOutVariable<string>? path)
         {
             Path = path;
         }
 
+        protected sealed override void initialize()
+        {
+            this.Require(Path, nameof(Path));
+        }
+
         protected sealed override bool run()
         {
-            var path = Path.GetValue();
+            var path = Path!.GetValue();
+            if (string.IsNullOrEmpty(path))
+            {
+                return false;
+            }
 
             File.Delete(path);
